Changed the type of the user id from integer to long.

+++ src/main/java/co/altruix/pcc/api/cdm/ImmediateSchedulingRequest.java
@@ -16,5 +16,5 @@ package co.altruix.pcc.api.cdm;
  *
  */
 public interface ImmediateSchedulingRequest extends PccMessage {
-    Integer getUserId();
+    Long getUserId();
 }
